Changed SolutionUtilities to require that a SolutionFile be passed and updated all callers to use this new requirement.
This comes at the cost of exposing more of the Microsoft.Build dependency to this tooling; however we already had a dependency on it, so I am OK at this time. The advantage of this is it will reduce the number of times we have to reparse the file. Long term if we feel the abstraction is too leaky we can look to encapsulate it, however YAGNI.

This change is being made to support more functionality that requires the Solution file in more places; therefore we want to reduce the number of times we have to go to disk to parse this information (again at the cost of more memory).

Added unit tests (SolutionUtilitiesTests) to validate the functionality of this change. In addition StyleCop was removed from the Unit Tests. While I still think StyleCop is good for production code (especially when you have multiple developers) attempting to configure it to the needs of the unit tests are also another example of YAGNI. It has been removed for now.