Adding support for inserting Solution Configuration section. After this commit you will no longer be required to open the solution file in Visual Studio to populate this information.
This task was much more straight forward than once assumed; it leverages the SolutionFile API which provides information about all the configurations in the current solution file.

The remainder of the changes were mostly around getting the code more unit testable to assist in validating this change. The biggest difference is the removal of hardcoding various new line characters; instead we now favor returning enumerations of lines which are assumed to be inserted in the order they were returned. This improved unit testability by allowing us to perform order-insensitive compares of the resulting solution files.